Remove router logic from server bundle and split client bundle in smaller chunks (#832)
* Change RouterClient to be a shared component to reduce server bundle size

* Change ServerStateProvider to be a shared component and avoid extra bundling in client build

* Update changeset
