--- packages/hydrogen/src/components/Router/index.tsx
@@ -1,5 +1,0 @@
-export {
-  Router,
-  useRouter,
-  useLocation,
-} from '../../foundation/Router/Router.client';
