+++ packages/hydrogen/src/foundation/ServerStateProvider/index.ts
@@ -1,8 +1,2 @@
-export {
-  ServerStateProvider,
-  ServerStateContext,
-} from './ServerStateProvider.client';
-export type {
-  ServerState,
-  ServerStateContextValue,
-} from './ServerStateProvider.client';
+export {ServerStateProvider, ServerStateContext} from './ServerStateProvider';
+export type {ServerState, ServerStateContextValue} from './ServerStateProvider';
