+++ packages/hydrogen/src/utilities/tests/shopifyMount.tsx
@@ -5,7 +5,7 @@ import {DEFAULT_LOCALE} from '../../foundation/constants';
 
 import {ShopifyConfig} from '../../types';
 import {ShopifyProvider} from '../../foundation/ShopifyProvider';
-import {Router} from '../../components/Router';
+import {Router} from '../../foundation/Router/Router';
 import {ServerState, ServerStateProvider} from '../../foundation';
 
 type SetServerState = React.Dispatch<React.SetStateAction<ServerState>>;
